+++ src/tool/Metatool.Tools.Software/Config.cs
@@ -6,6 +6,5 @@ namespace Metatool.Tools.Software
     public class Config
     {
         public string[] ConfigFolders { get; set; }
-        public int    Option2 { get; set; } = 5;
     }
 }