+++ src/tool/Metatool.Tools.Software/RunMode.cs
@@ -1,0 +1,9 @@
+﻿namespace Metatool.Tools.Software
+{
+    public enum RunMode
+    {
+        Inherit,
+        Admin,
+        User
+    }
+}